Refactor StatusMessage
