+++ packages/meditrak-app/app/widgets/StatusMessage/constants.js
@@ -1,0 +1,9 @@
+/**
+ * Tupaia MediTrak
+ * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
+ */
+
+export const PRIMARY = 'primary';
+export const SECONDARY = 'secondary';
+export const STATUS_MESSAGE_SUCCESS = 'SUCCESS';
+export const STATUS_MESSAGE_ERROR = 'ERROR';
