+++ packages/meditrak-app/app/widgets/StatusMessage/index.js
@@ -1,0 +1,7 @@
+/**
+ * Tupaia MediTrak
+ * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
+ */
+
+export { StatusMessage } from './StatusMessage';
+export { STATUS_MESSAGE_SUCCESS, STATUS_MESSAGE_ERROR } from './constants';
