+++ packages/meditrak-app/app/widgets/StatusMessage/styles.js
@@ -1,14 +1,10 @@
 /**
  * Tupaia MediTrak
- * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
+ * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
  */
 
-import React from 'react';
-import { StyleSheet, View, ViewPropTypes } from 'react-native';
-import PropTypes from 'prop-types';
-import Icon from 'react-native-vector-icons/Ionicons';
+import { StyleSheet } from 'react-native';
 
-import { Text } from './Text';
 import {
   getThemeColorOneFaded,
   BORDER_RADIUS,
@@ -16,23 +12,8 @@ import {
   THEME_COLOR_THREE,
   THEME_COLOR_FIVE,
   THEME_TEXT_COLOR_FOUR,
-} from '../globalStyles';
-
-export const STATUS_MESSAGE_SUCCESS = 'SUCCESS';
-export const STATUS_MESSAGE_ERROR = 'ERROR';
-const PRIMARY = 'primary';
-const SECONDARY = 'secondary';
-
-export const getMessageIconName = type => {
-  switch (type) {
-    case STATUS_MESSAGE_SUCCESS:
-      return 'checkmark-circle';
-
-    case STATUS_MESSAGE_ERROR:
-    default:
-      return 'warning';
-  }
-};
+} from '../../globalStyles';
+import { PRIMARY, SECONDARY, STATUS_MESSAGE_ERROR, STATUS_MESSAGE_SUCCESS } from './constants';
 
 export const getMessageStyle = type => {
   switch (type) {
@@ -45,7 +26,7 @@ export const getMessageStyle = type => {
   }
 };
 
-const getIconStyle = theme => {
+export const getIconStyle = theme => {
   switch (theme) {
     case SECONDARY:
       return iconStyles[SECONDARY];
@@ -56,7 +37,7 @@ const getIconStyle = theme => {
   }
 };
 
-const getTextStyles = theme => {
+export const getTextStyles = theme => {
   switch (theme) {
     case SECONDARY:
       return textStyles[SECONDARY];
@@ -67,28 +48,7 @@ const getTextStyles = theme => {
   }
 };
 
-export const StatusMessage = ({ message, style, type, theme }) => (
-  <View style={[localStyles.message, getMessageStyle(type), style]}>
-    <Icon name={getMessageIconName(type)} style={getIconStyle(theme)} />
-    <Text style={getTextStyles(theme)}>{message}</Text>
-  </View>
-);
-
-StatusMessage.propTypes = {
-  message: PropTypes.string,
-  type: PropTypes.oneOf([STATUS_MESSAGE_ERROR, STATUS_MESSAGE_SUCCESS]),
-  style: ViewPropTypes.style,
-  theme: PropTypes.string,
-};
-
-StatusMessage.defaultProps = {
-  message: '',
-  type: STATUS_MESSAGE_ERROR,
-  style: null,
-  theme: PRIMARY,
-};
-
-const localStyles = StyleSheet.create({
+export const localStyles = StyleSheet.create({
   message: {
     padding: 8,
     marginBottom: 20,
@@ -107,7 +67,7 @@ const localStyles = StyleSheet.create({
 });
 
 const baseIconStyles = { marginRight: 10, marginVertical: 3 };
-const iconStyles = StyleSheet.create({
+export const iconStyles = StyleSheet.create({
   [PRIMARY]: {
     ...baseIconStyles,
     fontSize: 14,
@@ -117,7 +77,7 @@ const iconStyles = StyleSheet.create({
 });
 
 const baseTextStyles = { flex: 1 };
-const textStyles = StyleSheet.create({
+export const textStyles = StyleSheet.create({
   [PRIMARY]: { ...baseTextStyles, color: THEME_COLOR_ONE },
   [SECONDARY]: { ...baseTextStyles, color: THEME_TEXT_COLOR_FOUR, fontWeight: '500' },
 });
